+++ src/FeedbackProblem.ce.vue
@@ -83,11 +83,11 @@
 </template>
 
 <script setup>
-import CheckboxItem from "./shared/CheckboxItem.vue";
-import TextareaInput from "./shared/TextareaInput.vue";
-import ErrorSummary from "./shared/ErrorSummary.vue";
-import { Reasons, Answer } from "../const";
-import { setupValidation } from "../validation";
+import CheckboxItem from "./components/CheckboxItem.vue";
+import TextareaInput from "./components/TextareaInput.vue";
+import ErrorSummary from "./components/ErrorSummary.vue";
+import { Reasons, Answer } from "./constants/responses";
+import { setupValidation } from "./validation";
 import axios from "axios";
 import { setLocale } from "@vee-validate/i18n";
 import { useForm, useFormValues } from 'vee-validate';
@@ -100,7 +100,7 @@ setLocale(locale.value);
 
 const props = defineProps({
   contentId: {
-    Number,
+    type: Number,
     default: null,
   },
   versionId: {
