--- src/const.js
@@ -1,9 +1,0 @@
-export const Reasons = {
-  OTHER: "other"
-}
-
-export const Answer = {
-  NOT_ANSWERED: 0,
-  YES: 1,
-  NO: 2
-}
