+++ src/constants/responses.js
@@ -1,0 +1,9 @@
+export const Reasons = {
+    OTHER: "other"
+}
+
+export const Answer = {
+    NOT_ANSWERED: 0,
+    YES: 1,
+    NO: 2
+}
